requirements.txt
attrs==19.3.0
aws-cdk.assets==1.46.0
aws-cdk.aws-autoscaling==1.46.0
aws-cdk.aws-autoscaling-common==1.46.0
aws-cdk.aws-certificatemanager==1.46.0
aws-cdk.aws-cloudformation==1.46.0
aws-cdk.aws-cloudwatch==1.46.0
aws-cdk.aws-ec2==1.46.0
aws-cdk.aws-eks==1.46.0
aws-cdk.aws-elasticloadbalancing==1.46.0
aws-cdk.aws-elasticloadbalancingv2==1.46.0
aws-cdk.aws-events==1.46.0
aws-cdk.aws-iam==1.46.0
aws-cdk.aws-kms==1.46.0
aws-cdk.aws-lambda==1.46.0
aws-cdk.aws-logs==1.46.0
aws-cdk.aws-route53==1.46.0
aws-cdk.aws-s3==1.46.0
aws-cdk.aws-s3-assets==1.46.0
aws-cdk.aws-sns==1.46.0
aws-cdk.aws-sqs==1.46.0
aws-cdk.aws-ssm==1.46.0
aws-cdk.cdk-assets-schema==1.46.0
aws-cdk.cloud-assembly-schema==1.46.0
aws-cdk.core==1.46.0
aws-cdk.custom-resources==1.46.0
aws-cdk.cx-api==1.46.0
aws-cdk.region-info==1.46.0
cattrs==1.0.0
constructs==3.0.3
jsii==1.7.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
